--- projects/mc_interface/test/test_motor_relay.c
@@ -1,105 +1,0 @@
-#include "can.h"
-#include "can_transmit.h"
-#include "delay.h"
-#include "event_queue.h"
-#include "exported_enums.h"
-#include "gpio.h"
-#include "interrupt.h"
-#include "log.h"
-#include "motor_relay.h"
-#include "ms_test_helpers.h"
-#include "test_helpers.h"
-#include "unity.h"
-
-#define TEST_MOTOR_RELAY_NUM_RX_HANDLERS 5
-#define TEST_MOTOR_RELAY_CAN_DEVICE_ID 0x1
-
-#define TEST_MOTOR_RELAY_LEFT \
-  { .port = GPIO_PORT_B, .pin = 3 }
-#define TEST_MOTOR_RELAY_RIGHT \
-  { .port = GPIO_PORT_B, .pin = 9 }
-#define TEST_MOTOR_RELAY_DELAY_MS 10
-
-static CANStorage s_can;
-static CANRxHandler s_rx_handlers[TEST_MOTOR_RELAY_NUM_RX_HANDLERS];
-static MotorRelayStorage s_motor_relay;
-
-typedef enum {
-  TEST_MOTOR_RELAY_EVENT_CAN_RX = 0,
-  TEST_MOTOR_RELAY_EVENT_CAN_TX,
-  TEST_MOTOR_RELAY_EVENT_CAN_FAULT,
-} TestMotorRelayEvent;
-
-static StatusCode prv_ack_cb(CANMessageID msg_id, uint16_t device, CANAckStatus status,
-                             uint16_t num_remaining, void *context) {
-  CANAckStatus *ack_status = context;
-  *ack_status = status;
-
-  LOG_DEBUG("CAN ACK status %d\n", status);
-
-  return STATUS_CODE_OK;
-}
-
-void setup_test(void) {
-  event_queue_init();
-  interrupt_init();
-  soft_timer_init();
-
-  CANSettings can_settings = {
-    .device_id = TEST_MOTOR_RELAY_CAN_DEVICE_ID,
-    .bitrate = CAN_HW_BITRATE_500KBPS,
-    .rx_event = TEST_MOTOR_RELAY_EVENT_CAN_RX,
-    .tx_event = TEST_MOTOR_RELAY_EVENT_CAN_TX,
-    .fault_event = TEST_MOTOR_RELAY_EVENT_CAN_FAULT,
-    .tx = { GPIO_PORT_A, 12 },
-    .rx = { GPIO_PORT_A, 11 },
-    .loopback = true,
-  };
-  TEST_ASSERT_OK(can_init(&can_settings, &s_can, s_rx_handlers, TEST_MOTOR_RELAY_NUM_RX_HANDLERS));
-
-  MotorRelaySettings relay_settings = {
-    .left_relay = TEST_MOTOR_RELAY_LEFT,
-    .right_relay = TEST_MOTOR_RELAY_RIGHT,
-    .delay_ms = TEST_MOTOR_RELAY_DELAY_MS,
-  };
-  TEST_ASSERT_OK(motor_relay_init(&s_motor_relay, &relay_settings));
-}
-
-void teardown_test(void) {}
-
-void test_motor_relay_basic(void) {
-  // Ask to close the relay
-  volatile CANAckStatus status = NUM_STATUS_CODES;
-  CANAckRequest ack_request = {
-    .callback = prv_ack_cb,
-    .context = &status,
-    .expected_bitset = CAN_ACK_EXPECTED_DEVICES(TEST_MOTOR_RELAY_CAN_DEVICE_ID),
-  };
-  CAN_TRANSMIT_MOTOR_RELAY(&ack_request, EE_RELAY_STATE_CLOSE);
-
-  MS_TEST_HELPER_CAN_TX_RX_WITH_ACK(TEST_MOTOR_RELAY_EVENT_CAN_TX, TEST_MOTOR_RELAY_EVENT_CAN_RX);
-  TEST_ASSERT_OK(status);
-
-  GPIOAddress left_relay = TEST_MOTOR_RELAY_LEFT;
-  GPIOAddress right_relay = TEST_MOTOR_RELAY_RIGHT;
-
-  // Make sure that both relays are now closed. We allow some delay before checking for sequencing.
-  delay_ms(TEST_MOTOR_RELAY_DELAY_MS);
-  GPIOState state = NUM_GPIO_STATES;
-  gpio_get_state(&left_relay, &state);
-  TEST_ASSERT_EQUAL(GPIO_STATE_HIGH, state);
-  gpio_get_state(&right_relay, &state);
-  TEST_ASSERT_EQUAL(GPIO_STATE_HIGH, state);
-
-  // Try opening the relays
-  CAN_TRANSMIT_MOTOR_RELAY(&ack_request, EE_RELAY_STATE_OPEN);
-  MS_TEST_HELPER_CAN_TX_RX_WITH_ACK(TEST_MOTOR_RELAY_EVENT_CAN_TX, TEST_MOTOR_RELAY_EVENT_CAN_RX);
-  TEST_ASSERT_OK(status);
-
-  // Opening the relays does not require sequencing, so don't delay
-  state = NUM_GPIO_STATES;
-  gpio_get_state(&left_relay, &state);
-  TEST_ASSERT_EQUAL(GPIO_STATE_LOW, state);
-  gpio_get_state(&right_relay, &state);
-  TEST_ASSERT_EQUAL(GPIO_STATE_LOW, state);
-}
